requirements.txt
imutils==0.5.4
numpy==1.21.0
opencv-python==4.5.2.54
paho-mqtt==1.5.1
